requirements.txt
absl-py==1.4.0
astor==0.8.1
attrs==22.1.0
autopep8==2.0.2
bidict==0.22.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
cmake==3.26.0
connexion==2.14.1
contourpy==1.0.7
cycler==0.11.0
dlib==19.24.1
dnspython==2.3.0
Flask==2.2.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.2
fonttools==4.39.0
gast==0.5.3
greenlet==2.0.2
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
imutils==0.5.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jose==1.0.0
jsonschema==4.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
matplotlib==3.7.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.1
pycodestyle==2.10.0
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
simple-websocket==0.9.0
six==1.16.0
SQLAlchemy==1.4.42
swagger-ui-bundle==0.0.9
tensorboard==1.12.2
termcolor==2.2.0
threadpoolctl==3.1.0
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.3
wsproto==1.2.0
yapf==0.32.0
zipp==3.15.0
